+++ R/overimp2D.R
@@ -1,12 +1,12 @@
 # overimp2D_scatter ------------------------------------------------------------
-overimp2D_scatter <- function(plot_data, x, y, title, subtitle, alpha, point_size, xlim, ylim, train_color_pal, test_color_pal) {
+overimp2D_scatter <- function(plot_data, x, y, title, subtitle, alpha, point_size, xlim, ylim, train_color_pal, test_color_pal, gg_style = list()) {
   train_plot <- ggplot(plot_data$train$all_dt, aes(x = .data[[x]], y = .data[[y]], color = Group)) +
     geom_point(alpha = alpha, size = point_size) +
     facet_grid(rows = vars(Group)) +
     scale_color_manual(values = train_color_pal) +
     labs(title = "Training Data", x = x, y = y) +
     guides(color = "none")
-  train_plot <- .ggplot_overimp_theme(train_plot, showlegend = FALSE)
+  train_plot <- .ggplot_overimp_theme(train_plot, showlegend = FALSE, gg_style = gg_style)
 
   if (!is.null(xlim)) train_plot <- train_plot + xlim(xlim)
   if (!is.null(ylim)) train_plot <- train_plot + ylim(ylim)
@@ -22,7 +22,7 @@ overimp2D_scatter <- function(plot_data, x, y, title, subtitle, alpha, point_siz
       scale_color_manual(values = test_color_pal) +
       labs(title = "Test Data", x = x, y = y) +
       guides(color = "none")
-    test_plot <- .ggplot_overimp_theme(test_plot, showlegend = FALSE)
+    test_plot <- .ggplot_overimp_theme(test_plot, showlegend = FALSE, gg_style = gg_style)
 
     if (!is.null(xlim)) test_plot <- test_plot + xlim(xlim)
     if (!is.null(ylim)) test_plot <- test_plot + ylim(ylim)
@@ -49,7 +49,7 @@ overimp2D_scatter <- function(plot_data, x, y, title, subtitle, alpha, point_siz
 
 
 # overimp2D_box ----------------------------------------------------------------
-overimp2D_box <- function(plot_data, x, y, title, subtitle, alpha, point_size, boxpoints, xlim, ylim, train_color_pal, test_color_pal) {
+overimp2D_box <- function(plot_data, x, y, title, subtitle, alpha, point_size, boxpoints, xlim, ylim, train_color_pal, test_color_pal, gg_style = list()) {
   train_plot <- ggplot(plot_data$train$all_dt, aes(x = .data[[x]], y = .data[[y]])) +
     facet_grid(rows = vars(Group)) +
     scale_fill_manual(values = train_color_pal) +
@@ -68,7 +68,7 @@ overimp2D_box <- function(plot_data, x, y, title, subtitle, alpha, point_size, b
     train_plot <- train_plot +
       geom_boxplot(alpha = alpha, aes(fill = Group), outlier.size = point_size, outlier.alpha = alpha)
   }
-  train_plot <- .ggplot_overimp_theme(train_plot, showlegend = FALSE)
+  train_plot <- .ggplot_overimp_theme(train_plot, showlegend = FALSE, gg_style = gg_style)
 
   if (!is.null(xlim)) train_plot <- train_plot + xlim(xlim)
   if (!is.null(ylim)) train_plot <- train_plot + ylim(ylim)
@@ -94,7 +94,7 @@ overimp2D_box <- function(plot_data, x, y, title, subtitle, alpha, point_size, b
       test_plot <- test_plot +
         geom_boxplot(alpha = alpha, aes(fill = Group), outlier.size = point_size, outlier.alpha = alpha)
     }
-    test_plot <- .ggplot_overimp_theme(test_plot, showlegend = FALSE)
+    test_plot <- .ggplot_overimp_theme(test_plot, showlegend = FALSE, gg_style = gg_style)
 
     if (!is.null(xlim)) test_plot <- test_plot + xlim(xlim)
     if (!is.null(ylim)) test_plot <- test_plot + ylim(ylim)
@@ -121,7 +121,7 @@ overimp2D_box <- function(plot_data, x, y, title, subtitle, alpha, point_size, b
 
 
 # overimp2D_bar ----------------------------------------------------------------
-overimp2D_bar <- function(plot_data, x, y, title, subtitle, width, alpha, ylim, train_color_pal, test_color_pal) {
+overimp2D_bar <- function(plot_data, x, y, title, subtitle, width, alpha, ylim, train_color_pal, test_color_pal, gg_style = list()) {
   train_plot <- ggplot(plot_data$train$all_dt, aes(x = .data[[x]], alpha = .data[[y]])) +
     geom_bar(aes(fill = Group, color = Group), width = width, position = "dodge2") +
     facet_grid(rows = vars(Group)) +
@@ -130,7 +130,7 @@ overimp2D_bar <- function(plot_data, x, y, title, subtitle, width, alpha, ylim, 
     scale_fill_manual(values = train_color_pal) +
     labs(title = "Training Data", x = x, y = y) +
     guides(fill = "none", color = "none")
-  train_plot <- .ggplot_overimp_theme(train_plot, showlegend = FALSE)
+  train_plot <- .ggplot_overimp_theme(train_plot, showlegend = FALSE, gg_style = gg_style)
 
   if (!is.null(ylim)) train_plot <- train_plot + ylim(ylim)
 
@@ -144,7 +144,7 @@ overimp2D_bar <- function(plot_data, x, y, title, subtitle, width, alpha, ylim, 
       scale_fill_manual(values = test_color_pal) +
       labs(title = "Test Data", x = x, y = y) +
       guides(fill = "none", color = "none", alpha = "none")
-    test_plot <- .ggplot_overimp_theme(test_plot, showlegend = FALSE)
+    test_plot <- .ggplot_overimp_theme(test_plot, showlegend = FALSE, gg_style = gg_style)
 
 
     if (!is.null(ylim)) test_plot <- test_plot + ylim(ylim)